Make the Rust daemon CI green (fmt, clippy, full test suite)
The published daemon crate was not self-contained for CI:

- fmt: apply `cargo fmt --all` across the workspace so the advisory fmt job
  is clean.
- clippy: resolve `-D warnings` — convert infallible `CString::new` unwraps in
  the PID-1 init helper to `.expect` under a justified crate allow, mark the
  provably-in-bounds u8-slot indexing in the work-history ring, add targeted
  allows for intentional patterns (NaN-fail-closed comparison, same-result
  clamp arms, multi-map index loop), replace a clone with `slice::from_ref`,
  and relax the two purely-cosmetic doc-formatting lints that the release
  scrub trips when it rewrites doc prose.
- full test suite: ship the `include_str!` data files the daemon embeds
  (docs/architecture/{install_matrix.tsv,hardware_enablement_matrix.json,
  artifact_producers.json}) so the tests compile, and drop the two tests that
  embed internal `.25` bring-up scripts the public export scrubs. The suite is
  advisory; the residual config-validation tests that assert on scrubbed
  worker fields remain non-blocking by design.

daemon-check (blocking) and dashboard remain green. No daemon behaviour
changed — only lint conformance, embedded data files, and doc prose.

+++ DCENT_OS_Antminer/dcentrald/dcentrald/src/work_dispatcher.rs
@@ -2374,7 +2374,7 @@ impl WorkDispatcher {
 
                             // Decode the nonce using chip-specific driver
                             let Some(decode_context) = FpgaNonceDecodeContext::try_new(
-                                chain.fpga_midstate_cnt as u8,
+                                chain.fpga_midstate_cnt,
                             ) else {
                                 hw_errors += 1;
                                 error!(
